Fix 'stop-on-error' bug #76
Since we added new sub commands to DocManager, we introduced
a new bug with the 'stop-on-error' function in XmlHandler.
For example: The analyze feature doesn't have a 'stop-on-error'
function. That means: Whenever we initialize a new XmlHandler
object for an XML file, the default value for 'stop-on-error'
is True. We have to make sure, that the argparser sets the
'stop-on-error' attribute by default to False.

This fixes #76